Moved new files to meet directory
