+++ react/features/base/meet/views/Home/components/PasswordInput.tsx
